Feat: páginas de Categorias e todos os produtos, sem paginação
