+++ frontend/src/components/ProductsPage/ProductsPage.jsx
@@ -1,7 +1,8 @@
-import { useContext, useEffect, useState } from "react";
-import { useNavigate, useParams } from "react-router-dom";
 import styled from "styled-components";
+import ProductThumb from "../Utils/ProductThumb";
+import { useContext, useEffect, useState } from "react";
 import UserContext from "../../contexts/UserContext";
+import { useNavigate, useParams } from "react-router-dom";
 
 const Wrapper = styled.div`
     display: flex;
@@ -14,39 +15,20 @@ const Wrapper = styled.div`
 
 const Header = styled.div`
     display: flex;
-    justify-content: space-between;
-    align-items: start;
+    justify-content: center;
+    align-items: center;
     padding: 0 25px;
-    position: absolute;
-    width: 100%;
+    position: relative;
     height: 70px;
+    width: 100%;
+    flex: 0 0 auto;
     z-index: 20;
-    top: 30px;
 `;
 
-const LikeButton = styled.div`
-    width: 56px;
-    height: 56px;
-    background-color: white;
-    border-radius: 100%;
-
-    display: flex;
-    align-items: center;
-    justify-content: center;
-
-    ion-icon {
-        font-size: 24px;
-        --ionicon-stroke-width: ${({ selected }) => (selected ? "0px" : "50px")};
-        color: ${({ selected, theme }) => {
-            return selected ? theme.special : theme.secondary;
-        }};
-    }
-    cursor: pointer;
-    :hover {
-        ion-icon {
-            color: ${({ theme }) => theme.special};
-        }
-    }
+const Title = styled.div`
+    font-size: 18px;
+    font-weight: bold;
+    text-align: center;
 `;
 
 const BackButton = styled.div`
@@ -54,10 +36,12 @@ const BackButton = styled.div`
     height: 56px;
     background-color: rgba(255, 255, 255, 0.2);
     border-radius: 100%;
+    left: 0px;
 
     display: flex;
     align-items: center;
     justify-content: center;
+    position: absolute;
 
     ion-icon {
         font-size: 24px;
@@ -70,183 +54,46 @@ const BackButton = styled.div`
     }
 `;
 
-const ImageArea = styled.div`
-    width: 100%;
-    flex: 1 1 auto;
-    overflow: hidden;
-
+const Products = styled.div`
     display: flex;
-    align-items: center;
+    flex-wrap: wrap;
     justify-content: center;
-
-    img {
-        transition: all 0.2s linear;
-        width: 100%;
-        height: 100%;
-        object-fit: cover;
-
-        :hover {
-            width: 110%;
-            height: 110%;
-        }
-    }
-`;
-
-const Bottom = styled.div`
-    display: flex;
-    flex-direction: column;
+    gap: 30px;
     width: 100%;
-    max-width: 500px;
-    height: 200px;
-    flex: 0 0 auto;
-    padding: 25px;
-    justify-content: space-between;
+    padding: 0px 25px 50px 25px;
+    flex: 1 1 auto;
+    overflow-y: auto;
+    align-content: start;
 `;
 
-const Texts = styled.div`
-    display: flex;
-    flex-direction: column;
-    width: 100%;
-    gap: 10px;
-`;
-
-const TextsTop = styled.div`
-    display: flex;
-    justify-content: space-between;
-    width: 100%;
-`;
-
-const Title = styled.h1`
-    font-weight: bold;
-    font-size: 24px;
-`;
-
-const Price = styled.h2`
-    font-weight: bold;
-    font-size: 36x;
-`;
-
-const Description = styled.p`
-    font-size: 14px;
-`;
-
-const Button = styled.button`
-    align-self: center;
-    border-radius: 12px;
-    border-style: none;
-    background-color: ${({ theme }) => theme.main};
-    :hover {
-        background-color: ${({ theme }) => theme.hoverMain};
-    }
-    color: ${({ theme }) => theme.overMain};
-    font-weight: 700;
-    font-size: 18px;
-    padding: 20px 0;
-    width: 100%;
-    cursor: pointer;
-`;
+const PageTitle = styled.div``;
 
 export default function ProductsPage() {
-    const { productId } = useParams();
+    const { categoryName } = useParams();
 
-    const { setCart, setLikes, getData, putData, cart, likes, token } = useContext(UserContext);
-
-    const [isLiked, setIsLiked] = useState(false);
-    const [isInCart, setIsInCart] = useState(false);
-    const [product, setProduct] = useState(null);
+    const [products, setProducts] = useState(null);
+    const [newProducts, setNewProducts] = useState([]);
+    const [lastPage, setLastPage] = useState(0);
+    const { getData } = useContext(UserContext);
 
     const navigate = useNavigate();
 
     useEffect(() => {
-        console.log(`producr!!! `, `product/${productId}`);
-        if (!product) {
-            getData(`/products/${productId}`, setProduct);
+        if (!products) {
+            getData(`/products/`, setProducts);
         }
-
-        if (product && likes) {
-            const likeProduct = likes.products.find((p) => {
-                return p.productId === product._id;
-            });
-            if (likeProduct) {
-                setIsLiked(true);
-            } else {
-                setIsLiked(false);
-            }
-        }
-        if (product && cart) {
-            const cartProduct = cart.find((p) => {
-                return p.productId === product._id;
-            });
-            if (cartProduct) {
-                setIsInCart(true);
-            } else {
-                setIsInCart(false);
-            }
-        }
-    }, [cart, likes, token, product]);
-
-    async function likeButtonPressed(event) {
-        event.stopPropagation();
-        if (!token) {
-            navigate("/signin");
-        }
-        try {
-            await putData("/wishlist", {
-                productId: product._id,
-                action: isLiked ? "delete" : "add",
-            });
-            setIsLiked(!isLiked);
-            await getData("/wishlist", setLikes);
-        } catch (err) {
-            return console.log(err.error);
-        }
-    }
-
-    async function cartButtonPressed(event) {
-        event.stopPropagation();
-        if (!token) {
-            navigate("/signin");
-        }
-        try {
-            const cartProduct = cart.find((p) => {
-                return p.productId === product._id;
-            });
-
-            if (isInCart) {
-                navigate("/cart");
-                return;
-            } else {
-                await putData("/cart", {
-                    productId: product._id,
-                    quantity: 1,
-                });
-                setIsInCart(!isInCart);
-                await getData("/cart", setCart);
-            }
-        } catch (err) {
-            return console.log(err.error);
-        }
-    }
+    }, []);
 
     async function backButtonPressed(event) {
         navigate(-1);
     }
 
-    const Image = product ? <img src={product.thumbnail}></img> : <></>;
-
-    const bottom = product ? (
-        <>
-            <Texts>
-                <TextsTop>
-                    <Title>{product.title}</Title>
-                    <Price>{product.price}</Price>
-                </TextsTop>
-                <Description>{product.description}</Description>
-            </Texts>
-            <Button>
-                {token ? (isInCart ? "Ver carrinho!" : "Adicionar ao carrinho") : "Fazer login"}
-            </Button>
-        </>
+    const thumbs = products ? (
+        products.map((product, index) => {
+            return (
+                <ProductThumb key={index} product={product} showCartButton={true}></ProductThumb>
+            );
+        })
     ) : (
         <></>
     );
@@ -257,12 +104,9 @@ export default function ProductsPage() {
                 <BackButton onClick={backButtonPressed}>
                     <ion-icon name="chevron-back-outline"></ion-icon>
                 </BackButton>
-                <LikeButton selected={isLiked} onClick={likeButtonPressed}>
-                    <ion-icon name={`heart${isLiked ? "" : "-outline"}`}></ion-icon>
-                </LikeButton>
+                <Title>Todos os produtos</Title>
             </Header>
-            <ImageArea>{Image}</ImageArea>
-            <Bottom onClick={cartButtonPressed}>{bottom}</Bottom>
+            <Products>{thumbs}</Products>
         </Wrapper>
     );
 }
